requirements.txt
viam-sdk==0.38.0
torch
torchvision
pillow
pytest
